+++ app/utils/cron.py
@@ -2,6 +2,7 @@
 from datetime import datetime, time, date
 import pytz
 from app.models import VendorEngagement, Location, MealSession
+from app.business_logic.meal_session.meal_session_logic import MealSessionLogic
 
 
 class MealSessionCron(object):
@@ -38,8 +39,7 @@ class MealSessionCron(object):
 			}
 		}
 
-	@staticmethod
-	def _complete_meal_session_construction_and_save(meal_sessions, date_time_to_use):
+	def _complete_meal_session_construction_and_save(self, meal_sessions, date_time_to_use):
 		"""
 		Complete the contruction of the meal sessions
 		and save if they do not already exist
@@ -53,10 +53,30 @@ class MealSessionCron(object):
 								 day=date_time_to_use.day)
 		meal_sessions['breakfast']['date'] = meal_session_date
 		meal_sessions['lunch']['date'] = meal_session_date
+
 		for meal_session in meal_sessions.values():
-			if MealSession.query.filter_by(**meal_session).all():
+			# Ignore creation of meal session if one with in the specified
+			# time already exists
+			if MealSessionLogic.validate_meal_session_times(
+				**{
+					"name": meal_session.get('name'),
+					"date_sent": meal_session.get('date'),
+					"location_id": meal_session.get('location_id'),
+					"start_time": meal_session.get('start_time'),
+					"end_time": meal_session.get('stop_time'),
+				}
+			):
 				continue
+
 			MealSession(**meal_session).save()
+
+	@staticmethod
+	def _get_scheduler_current_date():
+		return datetime.now(tz=pytz.timezone("Africa/Lagos"))
+
+	@staticmethod
+	def _get_location_current_date(city):
+		return datetime.now(tz=pytz.timezone("Africa/" + city))
 
 	def job_to_schedule(self):
 		"""
@@ -64,11 +84,9 @@ class MealSessionCron(object):
 
 		:return (None):
 		"""
-		# def dynamic_callable():
 		with self.app.app_context():
-			scheduler_datetime = datetime.now(tz=pytz.timezone("Africa/Lagos"))
-			print("Scheduler date time ------>", str(scheduler_datetime))
-			# meal_sessions_with_locations = []
+			scheduler_datetime = self._get_scheduler_current_date()
+
 			locations = self._extract_current_locations()
 			location_names = {location.id: location.name for location in locations}
 			for location in locations:
@@ -76,13 +94,16 @@ class MealSessionCron(object):
 				meal_sessions['breakfast']['location_id'] = location.id
 				meal_sessions['lunch']['location_id'] = location.id
 				location_name = location_names.get(location.id)
-				location_datetime = datetime.now(tz=pytz.timezone("Africa/"+ location_name))
+				# location_datetime = datetime.now(tz=pytz.timezone("Africa/" + location_name))
+				location_datetime = self._get_location_current_date(location_name)
 				if any([scheduler_datetime.year > location_datetime.year,
 					   scheduler_datetime.month > location_datetime.month,
 					   scheduler_datetime.day > location_datetime.day]):
 					self._complete_meal_session_construction_and_save(meal_sessions, scheduler_datetime)
 				else:
 					self._complete_meal_session_construction_and_save(meal_sessions, location_datetime)
+
+			print("    *** Meal session Cron Completed last run at: ", str(scheduler_datetime))
 
 
 
@@ -94,6 +115,9 @@ class Cron:
 
 	def run_24_hourly(self):
 		self.update_engagement_status()
+
+	def run_meal_session_cron(self):
+		self.meal_session_cron.job_to_schedule()
 
 	def update_engagement_status(self):
 		'''A cron job that periodically update an engagement status'''
